refactor(dts): split declaration emit shards under cap
