+++ crates/tsz-emitter/src/declaration_emitter/helpers/mod.rs
@@ -290,6 +290,7 @@ mod type_inference_object_unions;
 mod type_inference_package_matching;
 mod type_inference_portable_mapped_objects;
 mod type_inference_public_packages;
+mod type_inference_return_guards;
 mod type_inference_return_normalization;
 mod type_inference_return_surface;
 mod type_inference_return_unions;
